Add files via upload
This project deals with the overall requirements for the “All-Together Converter” In One Frame.

# All-Together Converter

* A python GUI application which converts different quantities like Area, Length, Weight, Temperature and Currency.This All In One Converter App is in Python.

All In One Converter project is a desktop application which is developed in Python platform. This is desktop application 100% working smooth without any bug. It is developed using Python and Database No Database are used this project.The beat One converter App In Python is simple project developed using Python. This project is GUI application which converts different quantities like Area, Length, Weight, Temperature and Currency from one unit to a different . Also, this app is capable of handling all kinds of exceptions. This project is a stimulating useful project.
In this desktop application (project) user have to use Currency Converter,  Temperature Converter, Length Converter, Area Converter, Weight converter. This is Multipurpose  Desktop Application.

Talking about the features of this system, this python application is designed to convert entered numbers from one system/unit to other system/unit and it is also capable of handling all types of exceptions.

Tkinter()-It is a standard Python interface to the Tk GUI toolkit shipped with Python. Python with tkinter outputs the fastest and easiest way to create the GUI applications. Creating a GUI using tkinter is an easy task.

This python application is meant to convert entered numbers from one system/unit to other system/unit and it's also capable of handling all kinds of exceptions. Module Used – Tkinter()-It may be a standard Python interface to the Tk GUI toolkit shipped with Python. Python with tkinter outputs the fastest and easiest method to make the GUI applications. Creating a GUI using tkinter is a simple task. Also, the planning of this technique is pretty simple in order that the user won’t get any difficulties while performing on it

Module Used :

Currency Converter :- 
						A currency converter will help international travellers study the price of foreign goods and services, helping cost-conscious visitors find out how much their money is worth while travelling overseas.

Temperature Converter :- 
							    A temperature converter helps in the conversion of the measurement units of the temperature recorded in a particular unit. Temperature expresses the degree of heat or cold of a solid, liquid, or gas. Temperature is measured using a thermometer. While Kelvin (K) is the SI unit of temperature, people generally use Centigrade or Celsius (°C) and Fahrenheit (°F) to measure temperature

Length Converter :-
					  Length conversion is the process of converting lengths from one unit to another. For example, kilometers can be converted to meters, miles can be converted to yards, yards can be converted to inches, and so on. Length conversion is needed for matching measurements across different scales

Area Converter :- 
					   Use this area converter to convert instantly between acres, square centimeters, square feet, square miles and other metric and imperial area units


Weight Converter :-
						Weight converter for units including Kilograms, Pounds, Grams, Ounces, Stones etc. ... Use the search box to find your required metric converter
 calculator : simple calculation
Editor : like notepad